- uninitialised variable in reprobe for rtt blocked domains fixed. - lintfix and new flex output.
git-svn-id: file:///svn/unbound/trunk@2583 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ daemon/daemon.c
@@ -321,7 +321,7 @@ daemon_create_workers(struct daemon* daemon)
 		if(!daemon->rand)
 			fatal_exit("could not init random generator");
 	}
-	hash_set_raninit(ub_random(daemon->rand));
+	hash_set_raninit((uint32_t)ub_random(daemon->rand));
 	shufport = (int*)calloc(65536, sizeof(int));
 	if(!shufport)
 		fatal_exit("out of memory during daemon init");
